### 2021-08-06 #3587 @whedon
[PRE REVIEW]: mctreesearch4j: A Monte Carlo Tree Search Implementation for the JVM
**Submitting author:** @larkz (<a href="http://orcid.org/0000-0002-9375-1035">Larkin Liu</a>)
**Repository:** <a href="https://github.com/JunTaoLuo/mctreesearch4j" target ="_blank">https://github.com/JunTaoLuo/mctreesearch4j</a>
**Version:** v0.0.3
