### 2021-09-27 #3587 @slim-patchy
# Review checklist for @slim-patchy

- [x] As the reviewer I confirm that I have read the JOSS conflict of interest policy and that there are no conflicts of interest for me to review this work.

- [x] I confirm that I read and will adhere to the JOSS code of conduct.

- [x]  Repository: Is the source code for this software available at the repository url?
- [x]  License: Does the repository contain a plain-text LICENSE file with the contents of an OSI approved software license?
- [ ]  Version: Does the release version given match the GitHub release?
- [x]  Authorship: Has the submitting author (@larkz) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?

### Comments
I believe the work would be a good addition to the existing literature and `JVM` ecosystem. I concur with my fellow reviewer @TimKam that better TDD could be implemented for the sake of CI/CD. Besides this aspect, since Timotheus has already covered more of the technical side of the codebase, so I will let the paper writing itself be the focus of my review. I am of the opinion that in order to establish the significance of the publication perhaps more discussion surrounding its motivation is warranted, with specific examples to how it can be applied, especially in a commercial setting. For example, the MCTS algorithm could be used in the software for self-driving cars such as in Tesla's Autopilot. This could be one selling point cited. I also think it odd that the algorithm's use in AlphaGo, AlphaGo Zero, AlphaZero, etc. which illustrates how it can be utilised in software that plays (board) games, has not been mentioned. Also, maybe more theory underpinning the algorithm can be introduced, so as (1) to explain it in terms of exploration versus exploitation, (2) to compare its pros and cons, and (3) to suggest any future improvements and/or enhancements to the programming work. A final remark is that perhaps adding a few lines to explain how it can be installed using `gradle` such as the basic instructions along with example commands would be particularly helpful to the more novice user.
