### 2021-10-07 #3804 @whedon
[REVIEW]: mctreesearch4j: A Monte Carlo Tree Search Implementation for the JVM
**Submitting author:** @larkz (<a href="http://orcid.org/0000-0002-9375-1035">Larkin Liu</a>)
**Repository:** <a href="https://github.com/JunTaoLuo/mctreesearch4j" target ="_blank">https://github.com/JunTaoLuo/mctreesearch4j</a>
**Version:** 0.0.4
