### 2021-11-18 #3804 @TimKam
I added some additional comments to the PR (with comprehensive changes) made by the authors: https://github.com/aqtech-ca/mctreesearch4j/pull/68
