### 2021-11-28 #3804 @larkz
@TimKam @adi3 @slim-patchy all of the comments regarding our paper and code have been addressed (from the authors' perspective) - please see https://github.com/aqtech-ca/mctreesearch4j/pull/68, we would kindly ask for a resolution moving forward. And feel free to let us know if further editions are required.
