### 2021-12-27 #3804 @TimKam
Apologies for getting back to this late. I thank the authors for addressing my comments. As mentioned initially, I am somewhat on the fence w.r.t. *substantial scholarly effort* because the MC tree-search is AI 'from the textbook'. Still, there is probably some value in the work and I have enjoyed trying it out a bit. I have no further feedback or change requests.
