### 2021-12-28 #3804 @larkz
thanks @TimKam for you review notes. Much appreciated!

To reiterate our main goal from doing this project was
a) Provide a stable and proven implementation of MCTS for the JVM ecosystem.
b) Provide a ready software platform, where researchers can already have the tooling to experiment with new MCTS modifications.
c) Provide key software design principles, specifically for MCTS, where under such design principles, one can also build a modular and efficient MCTS software.

From our perspective, the level of scholarly effort is rather incremental, but nevertheless the goal was to achieve a) b) and c) rather than reinvent MCTS algorithmically (of which future researchers are more than free to do using our software).

Happy holidays and many thanks again for your suggestions. @adi3 please move this forward per your discretion.
