### 2022-01-10 #3804 @adi3
@TimKam - thank you for finishing the review! We greatly appreciate your effort and the diligence with which you proceeded.

@larkz - Thank you for your patience through the process. The goal of JOSS is to provide visibility to robust and efficient implementations of software that could be of help to other academics or the wider research community, even if the code itself does not push the needle in terms of inventing something new. In view of this, I will be happy to recommend your paper for publishing. Thanks again!
