### 2022-01-10 #3804 @larkz
@adi3 hi the changes have been merged and a release 0.0.4 corresponding to this PR has been made. We created a record on Zenodo with DOI 10.5281/zenodo.5834890 [link](https://zenodo.org/record/5834890#.Ydy5dOrMKrw)

Please let us know if anything else!
