### 2022-02-04 #3804 @Kevin-Mattheus-Moerman
@larkz I have read the paper and inspected the ZENODO archive. Below are some final points to work on.

- [ ] Consider explaining the acronym `JVM` in the summary (currently done in the next Statement of Need section)
- [ ] Check `...is able inherit...` in the conclusion section, this should probably read `...is able to inherit...`
- [ ] Please edit the ZENODO archive (you need to do this manually) such that the title and author fields fully match that of the paper. Currently the title does not match. You may also wish to add author ORCID information at this point, and to generally make the archive meta data complete e.g. with a description etc. but we leave this up to you.
- [ ] It looks like I should update the version tag to `v0.0.4`. I'll process this now but please confirm if this is correct.
