### 2022-02-04 #3804 @larkz
hello @Kevin-Mattheus-Moerman
I've addressed to two grammar/wording isseus.
To confirm, yes the release version is `0.0.4` you may change accordingly.

- To ask, what should the title of the Zenodo archive be? The authors match, but I looked at other JOSS papers and it seems like the pattern is {repo-name}:{some_comment}. Please let us know.
