### 2022-02-04 #3804 @larkz
@Kevin-Mattheus-Moerman unfortunately, I couldn't get the file to manually upload to the Zenodo archive, I think either there is a bug, or I'm not using it correctly. It is working on the sandbox, https://sandbox.zenodo.org/record/1007518#.Yf2TE-pKirw

I am just waiting until the bug/issue with Zenodo is resolved... https://github.com/zenodo/zenodo/issues/2290..., if you happen to know how to resolve this please let me know in advance.
