### 2022-02-04 #3804 @larkz
@Kevin-Mattheus-Moerman, the previous Zenodo issue seems to have resolved itself (just waited for ~1 hr), we put the article up with a new DOI https://zenodo.org/record/5975821. Please update accordingly,

I think all the action items from your side are complete, let me know if you require anything else.
